[ks49team01 - SHJ] 객실관리 detail: 컨트롤러,서비스,매퍼,dto 패키지에 담당 파일 room한개에서 여러개로 나눠서 생성
