+++ ks49team01/src/main/java/ks49team01/admin/dto/AdminRoom.java
@@ -1,5 +1,5 @@
 package ks49team01.admin.dto;
 
-public class AdminRoomInfo {
+public class AdminRoom {
 
 }
