+++ ks49team01/src/main/java/ks49team01/admin/mapper/AdminOptionMapper.java
@@ -3,6 +3,6 @@ package ks49team01.admin.mapper;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
-public interface AdminReservationMapper {
+public interface AdminOptionMapper {
 
 }
