+++ ks49team01/src/main/java/ks49team01/admin/mapper/AdminRoomMapper.java
@@ -1,0 +1,13 @@
+package ks49team01.admin.mapper;
+
+import java.util.List;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import ks49team01.admin.dto.AdminSurcharge;
+
+@Mapper
+public interface AdminRoomMapper {
+	
+	
+}
