+++ ks49team01/src/main/java/ks49team01/admin/service/AdminOptionService.java
@@ -9,6 +9,6 @@ import lombok.AllArgsConstructor;
 @Service
 @Transactional
 @AllArgsConstructor
-public class AdminReservationService {
+public class AdminOptionService {
 
 }
