+++ ks49team01/src/main/java/ks49team01/admin/service/AdminRoomService.java
@@ -6,23 +6,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import ks49team01.admin.dto.AdminSurcharge;
-import ks49team01.admin.mapper.RoomMapper;
+import ks49team01.admin.mapper.AdminRoomMapper;
 import lombok.AllArgsConstructor;
 
 @Service
 @Transactional
 @AllArgsConstructor
-public class RoomService {
+public class AdminRoomService {
 	
 	
-	private RoomMapper roomMapper;
-	
-	/**
-	 * 시즌추가금리스트조회
-	 * @return List<AdminSurchargeList>
-	 */
-	public List<AdminSurcharge> getAdminSurchargeList(){
-		
-		return roomMapper.getAdminSurchargeList();
-	};
 }
