+++ ks49team01/src/main/java/ks49team01/branch/mapper/BranchOptionMapper.java
@@ -3,6 +3,6 @@ package ks49team01.branch.mapper;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
-public interface BranchReservationMapper {
+public interface BranchOptionMapper {
 
 }
