+++ ks49team01/src/main/java/ks49team01/branch/service/BranchRoomService.java
@@ -1,0 +1,8 @@
+package ks49team01.branch.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BranchRoomService {
+
+}
