--- ks49team01/src/main/java/ks49team01/branch/service/RoomService.java
@@ -1,5 +1,0 @@
-package ks49team01.branch.service;
-
-public class RoomService {
-
-}
