+++ ks49team01/src/main/java/ks49team01/branch/service/branchOptionService.java
@@ -3,6 +3,6 @@ package ks49team01.branch.service;
 import org.springframework.stereotype.Service;
 
 @Service
-public class BranchReservationService {
+public class branchOptionService {
 
 }
