+++ ks49team01/src/main/java/ks49team01/user/service/UserRoomService.java
@@ -1,5 +1,5 @@
 package ks49team01.user.service;
 
-public class RoomService {
+public class UserRoomService {
 
 }
